@@ -6,42 +6,24 @@ import 'package:scroll_app_bar/scroll_app_bar.dart';
 void main() => runApp(App());
 
 class App extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) => MaterialApp(home: Home());
-}
-
-class Home extends StatefulWidget {
-  @override
-  _HomeState createState() => _HomeState();
-}
-
-class _HomeState extends State<Home> {
-  ScrollAppBarController controller;
-
-  @override
-  void initState() {
-    super.initState();
-    controller = ScrollAppBarController();
-  }
-
-  @override
-  void dispose() {
-    controller.dispose();
-    super.dispose();
-  }
+  final controller = ScrollController();
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: ScrollAppBar(
-        scrollAppBarController: controller,
-        title: Text("App Bar"),
+    return MaterialApp(
+      home: Scaffold(
+        appBar: ScrollAppBar(
+          controller: controller,
+          title: Text("App Bar"),
+        ),
+        body: Snap(
+          controller: controller.appBar,
+          child: ListView.builder(
+            controller: controller,
+            itemBuilder: _listBuildItem,
+          ),
+        ),
       ),
-      body: ListView.builder(
-        controller: controller.scrollController,
-        itemBuilder: _listBuildItem,
-      ),
-      floatingActionButton: _pin,
     );
   }
 
@@ -50,13 +32,6 @@ class _HomeState extends State<Home> {
       padding: const EdgeInsets.symmetric(vertical: 50),
       color: Color(Random().nextInt(0xffffffff)),
       child: Center(child: Text("$index")),
-    );
-  }
-
-  Widget get _pin {
-    return FloatingActionButton(
-      onPressed: () => controller.tooglePin(),
-      child: Icon(Icons.touch_app),
     );
   }
 }
