@@ -1,4 +1,6 @@
 library scroll_app_bar;
 
-export 'src/scroll_app_bar_controller.dart';
+export 'package:scroll_bars_common/scroll_bars_common.dart' show Snap;
+
 export 'src/scroll_app_bar.dart';
+export 'src/scroll_app_bar_controller.dart';
